Win / Lose
Moved the code for generating a new board into the generateBoard
function which can be called anytime we need a new board.

Hooked up the current level and highestlevel level to the save and load
feature.

Next two sections are in function checkWin():

Added logic to check if current score is >= levelScoreMultiplier *
currentLevel.  If it is then increases the currentLevel by 1 and
generates a new board and resets the scoreValue to 0.

Added logic to check if more than levelTime has passed and if so resets
the score and board and keeps currentLevel as is.