Updated header
Removed the graphics for the timer and the hops.  Moved the countdown
display to the right and added more space between the score and level
and highscore and highest level.  Adjusted where things land on the
page a bit too so it’s a little cleaner.  Need to make final
adjustments and also add that mute button!